some more major changes. api + db
