+++ src/lib/db/donationSettings.ts
@@ -1,40 +1,30 @@
-import { Donation_settings } from "@prisma/client";
+import { Donation_settings, Streamer } from "@prisma/client";
 import prisma from "../prisma";
 
-export const getDonationSettings = (
-  streamer: Donation_settings["streamer"]
+// TODO this is a public data point from every streamer
+// therefore, can we really find the steamer by ID?
+// maybe.
+
+export const getDonationSettings = async (
+  name: Streamer["name"]
 ) => {
   // TODO manual error handeling
-  return prisma.donation_settings.findUnique({
+
+  const streamer = await prisma.streamer.findUnique({
     where: {
-      streamer,
+      name,
     },
   });
+  if (streamer && streamer.id) {
+    const donationSettings = await prisma.donation_settings.findUnique({
+      where: {
+        streamer: streamer.id,
+      },
+    });
+    return donationSettings
+  }
+  return undefined
 };
-
-// export const removeAccount = (streamer: Account["streamer"]) => {
-//   return prisma.account.delete({
-//     where: {
-//       streamer,
-//     },
-//   });
-// };
-
-// export const createAccount = (
-//   streamer: Account["streamer"],
-//   data: {
-//     createdAt: Account["createdAt"];
-//     isOnline: Account["isOnline"];
-//     status: Statuses;
-//   }
-// ) => {
-//   return prisma?.account.create({
-//     data: {
-//       streamer,
-//       ...data,
-//     },
-//   });
-// };
 
 export const updateDonationSettings = (
   streamer: Donation_settings["streamer"],
