+++ src/pages/api/streamer/[id].ts
@@ -7,6 +7,7 @@ import {
 } from "~/lib/db/streamer";
 
 const handler: NextApiHandler = async (req, res) => {
+  console.log(req.query)
   switch (req.method) {
     case "GET":
       streamerGetHandler(req, res);
